ai-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pytesseract>=0.3.10
pdf2image>=1.16.0
pandas>=2.0.0
pillow>=10.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
numpy>=1.26.0
gunicorn>=21.0.0
